refactor: Tidy up the Psion Software Index views and models
This change strips out some unused code that was incorrectly brought into the project when folding-in the separate software index package. It also updates the license headers to apply the GPLv2.
